% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{receptoR
This stand-alone package allows for analysis of high-throughput transcriptome data, specifically identifying molecular receptors present on specific cell types of interest.}
\usage{
launchApp()
}
\value{
shiny application object
}
\description{
receptoR
This stand-alone package allows for analysis of high-throughput transcriptome data, specifically identifying molecular receptors present on specific cell types of interest.
}
